data-extract-index-API/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
langchain
azure-search-documents
openai
python-dotenv
langchain-community 
langchain-openai 
langchainhub 
tiktoken 
azure-ai-documentintelligence
azure-ai-formrecognizer

